fix: 항상 최단거리 기준으로 요금 산정하도록 수정

+++ src/main/java/nextstep/subway/domain/Price.java
@@ -27,6 +27,6 @@ public class Price {
     }
 
     private static int getOverFareSecond(double distance) {
-        return (int) Math.ceil((distance - 50) / 8) * 100;
+        return (int) Math.ceil((double) (distance - 50) / 8) * 100;
     }
 }
